requirements.txt
streamlit==1.31.0
openai==1.12.0
python-dotenv==1.0.1
requests==2.31.0
httpx==0.27.0
typing-extensions==4.9.0